Converting connection.js fie to ES6
